backend/foodgram/requirements.txt
Django==4.0.6
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-extra-fields==3.0.2
psycopg2==2.9.3
Pillow==9.2.0